Moved saving activity and it's routes logic from resolver to service.
